♻️  Refactoring the repository   - Updated structure of repository   - Renamed main package "ajpastor" -> "dd_functions"   - Removed folder for binder   - Updated VERSION to have better meaning   - Updated "requirements.txt"   - Moved notebooks to their folder
