--- binder/environment.yml
@@ -1,23 +1,0 @@
-name: sage-binder-env
-
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - jupyterlab
-  # Required by the SageMath kernel in Jupyter
-  - ipywidgets
-  # Install the full distribution of SageMath. The resulting image might take a
-  # long time to load on binder. Replace sage with sagelib, to use a minimal
-  # distribution of SageMath, without plotting, without symbolic computations, ….
-  - sage=9.6
-  # If you chose sagelib instead of sage above, you can add the required
-  # features manually here, e.g., matplotlib-base for basic plotting, sympy for
-  # symbolic computations, …. See the "run" section of
-  # https://github.com/conda-forge/sage-feedstock/blob/main/recipe/meta.yaml
-  # for the things that are included in "sage" but that might be missing from
-  # "sagelib".
-  - pip
-  - pip:
-    - git+https://github.com/mkauers/ore_algebra.git
-    - git+https://github.com/Antonio-JP/dd_functions.git
