+++ dd_functions/lazy/__init__.py
