+++ dd_functions/lazy/lazy/__init__.py
