+++ dd_functions/lazy/lazy/lazy_ring.py
