+++ dd_functions/misc/cached_property.py
