+++ dd_functions/misc/dynamic_string.py
