+++ dd_functions/misc/ring_w_sequence.py
