+++ dd_functions/operator/__init__.py
