+++ dd_functions/operator/listOperator.py
