+++ dd_functions/operator/oreOperator.py
