--- dependencies
@@ -1,5 +1,0 @@
-# no dependencies
-
-----------
-All lines of this file are ignored except the first.
-It is copied by SAGE_ROOT/build/make/install into SAGE_ROOT/build/make/Makefile.
