+++ requirements.txt
@@ -1,0 +1,2 @@
+ore_algebra @ git+https://github.com/mkauers/ore_algebra.git
+dalgebra @ git+https://github.com/Antonio-JP/dalgebra.git
