+++ setup.py
@@ -1,26 +1,25 @@
 ## -*- encoding: utf-8 -*-
-import os
 import sys
-from setuptools import setup
+from setuptools import setup, find_packages
 from codecs import open # To open the README file with proper encoding
 
 # Get information from separate files (README, VERSION)
 def readfile(filename):
     with open(filename,  encoding='utf-8') as f:
         return f.read()
-    
+
 setup(
-    name = "dd_functions",
+    name = "dd_functions", # The name of the package
     version = readfile("VERSION").strip(), # the VERSION file is shared with the documentation
-    description='A Sage package for DD-finite functions',
+    description='A Sage package for Differentially Definable functions',
     # long_description = readfile("README.txt"), # get the long description from the README
     # For a Markdown README replace the above line by the following two lines:
     long_description = readfile("README.md"),
     long_description_content_type="text/markdown",
-    url='https://github.com/Antonio-JP/dd_functions',
+    url='https://github.com/Antonio-JP/dd_functions', # URL to the project homepage
     author = "Antonio Jimenez-Pastor",
     author_email = "antonio.jimenezp@upm.es",
-    license = "GPLv3+", # See LICENCE file
+    license = "GPLv3+", # See LICENSE file
     classifiers=[
       # How mature is this project? Common values are
       #   3 - Alpha
@@ -31,15 +30,11 @@ setup(
       'Topic :: Software Development :: Build Tools',
       'Topic :: Scientific/Engineering :: Mathematics',
       'License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)',
-      'Programming Language :: Python :: 3.10.12',
+      f'Programming Language :: Python :: {sys.version}',
     ], # classifiers list: https://pypi.python.org/pypi?%3Aaction=list_classifiers
-    keywords = "holonomic SageMath dfinite",
-    packages = ["ajpastor",
-        "ajpastor.dd_functions",
-        "ajpastor.operator",
-        "ajpastor.lazy",
-        "ajpastor.misc"],
-    setup_requires   = ['sphinx'],
-    install_requires = ['ore_algebra @ git+https://github.com/mkauers/ore_algebra.git'],
+    keywords = "SageMath difference/differential algebra",
+    packages = find_packages(),
+    setup_requires   = [],
+    install_requires = readfile("requirements.txt").split("\n"),
 )
